refactor: extract OpenAPI config parsing into plenum-config crate (#141)
Move the config parsing, interpolation, and request context infrastructure
into a standalone crate that can be reused independently of plenum-core.

The new plenum-config crate provides:
- Template parsing and boot-time interpolation (${{ env.* }}, ${{ file.* }})
- Config struct with extension(), resolve(), and new extension_cascade()
- Runtime context extraction (ContextRef, ContextTemplate, ConfigValue)
- RequestData trait for HTTP framework decoupling

plenum-core now depends on plenum-config and re-exports all types,
keeping existing public API paths unchanged.
