+++ plenum-config/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "plenum-config"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+oas3 = { version = "0.20.1", features = ["yaml-spec"] }
+oapi-overlay = { version = "0.2.0", path = "../openapi-overlay", features = ["yaml"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+serde_yaml_ng = "0.9"
+form_urlencoded = "1"
