+++ plenum-core/src/load_balancing/pool.rs
@@ -9,7 +9,7 @@ use pingora_load_balancing::LoadBalancer;
 use pingora_load_balancing::selection;
 
 use crate::health_check::PassiveHealthTracker;
-use crate::request_context::{ContextRef, ExtractionCtx};
+use crate::request_context::{ContextRef, ExtractionCtx, PingoraRequest};
 
 /// Maps a resolved `SocketAddr` back to the original hostname configured for that
 /// backend. Used to set TLS SNI correctly (SNI must be the hostname, not the IP).
@@ -88,7 +88,7 @@ impl UpstreamPool {
         let key = match &self.hash_key_source {
             Some(ctx_ref) => ctx_ref
                 .extract(&ExtractionCtx {
-                    req,
+                    req: &PingoraRequest(req),
                     path_params,
                     user_ctx: None,
                     peer_addr: None,
